Replaced Stylix
This represents a massive undertaking for my own personal satisfaction.
Previously, danth/stylix, a module wrapping nixos module, was used extensively to rice my system apps.
This was easy to use, but caused a few issues (it is the reason I moved away from wezterm) and took agency away from me. Additionally, for programs that Stylix had not yet built ricing modules for, such as Zed, I was left without ability to rice it, leading me to actively avoid any program that did not have a module wrapper for Stylix.

The method as of this commit is somewhat unfinished, although it does work about how I wish it to work. I consider it a vertical slice.
I initially tried using nix-colors, however I found it more limiting compared to Stylix. So I wrote my own custom module, allowing me to define font size and allow me to extend it in the future as I see fit. I am likely to do so.

Overall, an excellent side-quest amidst the Hyprland switch.