hyprland: migrate majority over to hjem
this commit actually represents more work than it may seem. in order to cleanly migrate from home-manager, i needed to extend lib with the custom toHyprconf generator found in home-manager (this is, of course, licensed under MIT).
I also setup hyprland to run with UWSM to integrate the services better.
there is still some remnants of hyprland in hm, namely mako, swaylock, and hypridle.
this will be dealt with shortly.