Reorganize module system
with hm removed from my system, i am now free to reorganize the module system as i think makes the most sense.
previously, because hm and nix modules were incompatible with one another, it made the most sense to me to have three subdirectories in the module directory:
1. hm modules
2. nix modules
3. option declaration that is shared between the two types of modules.
because everything is now governed as a nix module, this kind of separation no longer makes sense, and i have made adjustments accordingly.
this has also allowed me to combine certain files that were related to the same kind of module but one was for hm and one was for nix.
for example, there is no longer a hyprland.nix completely isolated from my hyprland configuration; it has joined.
i am now contemplating what i will do with my options directory. i may integrate it or may not. this is also heavily dependent on the upcoming hjem-rum project.