flake: init flake-parts; lib: init listFilesRecursiveClean & importNixRecursive
Laying the groundwork for the move to a modular flake. My flake has gradually become a bit unworkable with astal, and as I would like to further expand my flake, a monolithic file has grown unwieldly and undesireable. This commit lays the ground work to begin modularizing my flake and also expands my extendedLib to modify a nixpkgs function for my imports.